requirements.txt
apache-airflow-client==2.9.0
google-api-python-client==2.141.0
praw==7.7.1
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
